Fixing a bug with image conversions.
It appears that popplerutils is suddenly padding zeroes on page counts.
This was an issue that I had before that caused me to create a hack to
get around the issue. Why it suddenly started doing what I wanted it to
do all along, I have no clue. As far as I know, I have not updated any
libraries. Very strange.